Move main application logic to library
